Separate loading slash commands from normal commands, split slash commands to separate cogs, make init_extensions support directories
